+++ kex_openapi/__init__.py
@@ -1,6 +1,6 @@
 """한국도로공사 OpenAPI 비공식 Python 클라이언트."""
 
-from pykrtour import PlaceCoordinate
+from pykrtour import AddressRegion, PlaceCoordinate
 
 from .client import KexClient
 from .codes import (
@@ -49,6 +49,7 @@ from .models import (
 )
 
 __all__ = [
+    "AddressRegion",
     "CarType",
     "CongestionLevel",
     "CoordinateSystem",